Crate model_generator

A library for programmatically generating 3D models through transformations and plugins.

This crate provides tools for creating and manipulating 3D models with a focus on composable transformations and an extensible plugin system.

Re-exports

pub use types::Vertex;

pub use types::Face;

pub use types::Mesh;

pub use transforms::basic::Scale;

pub use transforms::basic::Translate;

pub use transforms::basic::Rotate;

pub use transforms::advanced::Matrix;

pub use transforms::advanced::Mirror;

pub use transforms::advanced::Quaternion;

Modules

exporters

File format exporters for 3D models.

plugin

Plugin system for extending model-generator functionality.

primitives

Primitive 3D shapes that can be used as building blocks.

transforms

Transformations that can be applied to 3D models.

types

Core geometric types for the model-generator library.

Structs

Model

The core data structure representing a 3D model.

Enums

Error

Error types for the model-generator library.

Traits

Transform

Trait for implementing transformations that can be applied to a model.

Type Aliases

Result

Result type for operations in the model-generator library.